Fix a dangerous bug in lcm's marshalling and hash functions.

+++ lcm/ioutils.h
@@ -21,7 +21,10 @@ static inline int fwrite32(FILE *f, int32_t v)
 
 static inline int fwrite64(FILE *f, int64_t v64)
 {
-    int32_t v = v64>>32;
+    //  See Section 5.8 paragraph 3 of the standard
+    //  http://open-std.org/JTC1/SC22/WG21/docs/papers/2015/n4527.pdf
+    //  use uint for shifting instead if int
+    int32_t v = ((uint64_t)v64)>>32;
     if (0 != fwrite32(f, v)) return -1;
     v = v64 & 0xffffffff;
     return fwrite32(f, v);
@@ -49,7 +52,10 @@ static inline int fread64(FILE *f, int64_t *v64)
     if (fread32(f, &v2))
         return -1;
 
-    *v64 =     (((int64_t) v1)<<32) | (((int64_t) v2)&0xffffffff);
+    //  See Section 5.8 paragraph 3 of the standard
+    //  http://open-std.org/JTC1/SC22/WG21/docs/papers/2015/n4527.pdf
+    //  use uint for shifting instead if int
+    *v64 =     (int64_t)(((uint64_t) v1)<<32) | (((int64_t) v2)&0xffffffff);
 
     return 0;
 }
